# pkg/trace/agent: consider UTF-8 characters in tags