This change fixes a problem in NormalizeTag where it was assuming character lengths to be always equal to 1, whereas this isn't always true for UTF-8 characters.

It also fixes an issue where not all Unicode letters were lower-cased.

⚠️ **Warning**: there is an edge case to this algorithm where uppercased Unicode characters are not lowercased when their length would end up being different as a result. 24 characters fall into this edge case: https://play.golang.org/p/Af1bIusSbfq

Performance remains virtually unimpacted.

     name                      old time/op    new time/op    delta
     NormalizeTag/ok-4           68.8ns ± 0%    69.2ns ± 0%  +0.58%
     NormalizeTag/trim-4         92.8ns ± 0%    92.5ns ± 0%  -0.32%
     NormalizeTag/trim-both-4     159ns ± 0%     163ns ± 0%  +2.52%
     NormalizeTag/plenty-4        147ns ± 0%     150ns ± 0%  +2.04%
     NormalizeTag/more-4          259ns ± 0%     270ns ± 0%  +4.25%

     name                      old alloc/op   new alloc/op   delta
     NormalizeTag/ok-4            24.0B ± 0%     24.0B ± 0%   0.00%
     NormalizeTag/trim-4          32.0B ± 0%     32.0B ± 0%   0.00%
     NormalizeTag/trim-both-4     64.0B ± 0%     64.0B ± 0%   0.00%
     NormalizeTag/plenty-4        64.0B ± 0%     64.0B ± 0%   0.00%
     NormalizeTag/more-4           128B ± 0%      128B ± 0%   0.00%

     name                      old allocs/op  new allocs/op  delta
     NormalizeTag/ok-4             2.00 ± 0%      2.00 ± 0%   0.00%
     NormalizeTag/trim-4           2.00 ± 0%      2.00 ± 0%   0.00%
     NormalizeTag/trim-both-4      3.00 ± 0%      3.00 ± 0%   0.00%
     NormalizeTag/plenty-4         3.00 ± 0%      3.00 ± 0%   0.00%
     NormalizeTag/more-4           4.00 ± 0%      4.00 ± 0%   0.00%

In addition to https://github.com/DataDog/datadog-agent/pull/2951