@gbbr 2019-01-29
@AlexJF PTAL. FYI 54 characters will meet the edge case: https://play.golang.org/p/47am5UrqrM_N